code review changes implemented
